+++ tests/testthat/test-plots_uk.R
@@ -1,4 +1,5 @@
 test_fun <- function(x, prefix, ...) {
+  skip_on_cran()
   images <- list.files(system.file("img", package = "rasterpic"),
     full.names = TRUE
   )